requirements.txt
numpy
opencv-python-headless
streamlit
Pillow
mediapipe